Update to latest CS 162
The latest publicly available iteration of CS 162 is Spring 2015, while the one displayed in the README is from 2010. This one goes through new and emerging trends such as IoT, Cloud Computing, Quantum Computing, which feed back well into the other courses here.

Also, according to an announcement on the Youtube playlist page, this will be the last iteration of the course whose lectures will be posted to Youtube, so this shouldn't change for a while.