Add Probability Alternative
Added an alternative Probability course. It covers the same content as the existing one. Some would say that this one is better, the only reason I would recommend this one is that it has a full version outside of EdX, so learners can take their time to understand material if they choose not to pay. This one has many exercices for practice. It also has an EdX version, I tried it and didn't like it that much. **Optional book** not realted to this course but it covers everything literally step-by-step (it saved me in combinatronics): https://www.goodreads.com/en/book/show/29991429-probability