Adding OpenDSA interactive book
OpenDSA is an OSS project which offers an interactive web-based data structure and algorithms eTextbook
https://github.com/OpenDSA/OpenDSA/

I think it is a great DSA textbook which I used to read while studying the data structure course in the university. It is rich with animated examples which gives a great studying experience (almost every algorithm and data structure is animated and can be adjusted by the reader). 
I think it can be listed, as a DSA reference.