Update edx links in extra courses (#1343)
Most links were update to follow redirects.
The link to Delft University's Introduction to Functional Programming 404's, however, old links redirect to this URL, and so do search results on Edx. I assume this is a temporary issue from Edx and the page is supposed to be at this URL.

Colombia's Machine Learning and Data Science course is gone entirely, so that link was deleted.